+++ packages/conc/test/Polysemy/Conc/Test/SyncTest.hs
@@ -1,12 +1,17 @@
 module Polysemy.Conc.Test.SyncTest where
 
+import Data.Time (Day, UTCTime)
 import Polysemy.Test (UnitTest, assertEq, runTestAuto)
+import qualified Polysemy.Time as Time
+import Polysemy.Time (MicroSeconds (MicroSeconds), interpretTimeGhc)
 
+import Polysemy.Conc.AtomicState (interpretAtomic)
 import Polysemy.Conc.Effect.Race (Race)
 import qualified Polysemy.Conc.Effect.Sync as Sync
 import Polysemy.Conc.Effect.Sync (Sync)
 import Polysemy.Conc.Interpreter.Race (interpretRace)
-import Polysemy.Conc.Interpreter.Sync (interpretSync)
+import Polysemy.Conc.Interpreter.Sync (interpretSync, interpretSyncAs)
+import Polysemy.Conc.Sync (lock)
 
 data Thread1 = Thread1
 data Thread2 = Thread2
@@ -51,3 +56,18 @@ test_sync =
   runTestAuto do
     result <- run $ sequenceConcurrently @[] [thread1, thread2]
     assertEq @_ @IO [Just 3, Just 2] result
+
+test_syncLock :: UnitTest
+test_syncLock =
+  runTestAuto $
+  interpretTimeGhc $
+  asyncToIOFinal $
+  interpretRace $
+  interpretSyncAs () $
+  interpretAtomic (0 :: Int) do
+    void $ sequenceConcurrently @[] $ [(1 :: Int)..100] $> do
+      lock () do
+        cur <- atomicGet @Int
+        Time.sleep @UTCTime @Day (MicroSeconds 100)
+        atomicPut (cur + 1)
+    assertEq @Int @IO 100 =<< atomicGet